% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortree.R
\docType{package}
\name{colortree}
\alias{colortree}
\alias{colortree-package}
\title{colortree: Read colored nexus files and convert them to PDFs}
\description{
colortree: Read colored nexus files and convert them to PDFs
}
