% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortree.R
\name{default_export_color_nexus}
\alias{default_export_color_nexus}
\title{Export a plot to a PDF}
\usage{
default_export_color_nexus(g, path, width = 5, height = 5,
  units = "in", device = "pdf")
}
\arguments{
\item{g}{ggplot object}

\item{path}{filename for the output file}

\item{width}{width of the figure}

\item{height}{height of the figure}

\item{units}{units for width and height}

\item{device}{format of output (default="pdf")}
}
\value{
nothing
}
\description{
Export a plot to a PDF
}
