% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortree.R
\name{default_nexus2pdf}
\alias{default_nexus2pdf}
\title{Convert a nexus file to a pdf}
\usage{
default_nexus2pdf(trefile)
}
\arguments{
\item{trefile}{the path to a nexus file}
}
\description{
Convert a nexus file to a pdf
}
