% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortree.R
\name{default_plot_color_nexus}
\alias{default_plot_color_nexus}
\title{Plot the output of \code{read_color_nexus}}
\usage{
default_plot_color_nexus(tre)
}
\arguments{
\item{tre}{treedata object with data fields as specified in \code{read_color_nexus}}
}
\value{
ggplot object
}
\description{
Plot the output of \code{read_color_nexus}
}
