% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colortree.R
\name{read_color_nexus}
\alias{read_color_nexus}
\title{Read a nexus file with color annotations into a treedata object with a color column}
\usage{
read_color_nexus(trefile)
}
\arguments{
\item{trefile}{path to a nexus file}
}
\description{
Read a nexus file with color annotations into a treedata object with a color column
}
